Add "Add Product" UI and fix various bugs

+++ src/main/java/io/equalink/pricekeep/service/dto/ProductInfo.java
@@ -39,7 +39,7 @@ public record ProductInfo(
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
     Map<String, BigDecimal> stats,
 
-    Integer quantityPerItem,
+    BigDecimal quantityPerItem,
 
     Integer itemPerBundle,
